Created tile class
The tile class maintains information about the tiles of the board. The x,y location on the board, the neighbor locations, the state of the tile, (0, 1, 2), depending on whether the tile is empty, or occupied by player 1 or player 2, and edge_tile which is True if the tile is on the edge of the board.

Generate neighbors method populates the neighbors list. There are also some getters/setters for the other variables.